refactor(): 요청에 썸네일 이미지가 없을 시 기본 이미지를 저장 (#202)
* chore(): 파일명 확장자 구분 `.`으로 변경 및 파일명 생성 로직 분리

* refactor(): 요청에 썸네일 이미지가 없을 시 기본 이미지를 저장

* refactor(): 파일 확장자 기호 상수로 빼기

+++ backend/src/main/java/com/wooteco/nolto/image/application/ImageService.java
@@ -12,17 +12,22 @@ import javax.transaction.Transactional;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.Objects;
 
 @Transactional
 @Service
 public class ImageService {
-    private String defaultImageUrl = "https://dksykemwl00pf.cloudfront.net/nolto-default-thumbnail.png";
+
+    public static final String FILENAME_EXTENSION_DOT = ".";
 
     @Value("${application.bucket.name}")
     private String bucketName;
 
     @Value("${application.cloudfront.url}")
     private String cloudfrontUrl;
+
+    @Value("${application.default-image}")
+    private String defaultImage;
 
     private final AmazonS3 amazonS3Client;
 
@@ -31,15 +36,25 @@ public class ImageService {
     }
 
     public String upload(MultipartFile multipartFile) {
+        if (multipartFile.isEmpty()) {
+            return cloudfrontUrl + "/" + defaultImage;
+        }
         File file = convertToFile(multipartFile);
-        String fileName = System.currentTimeMillis() + Base64.encodeAsString(multipartFile.getName().getBytes()) + multipartFile.getContentType();
+        String fileName = getFileName(file);
         amazonS3Client.putObject(new PutObjectRequest(bucketName, fileName, file));
         file.delete();
         return cloudfrontUrl + "/" + fileName;
     }
 
+    private String getFileName(File file) {
+        String fileOriginName = file.getName();
+        int pos = fileOriginName.lastIndexOf(FILENAME_EXTENSION_DOT);
+        String ext = FILENAME_EXTENSION_DOT + fileOriginName.substring(pos + 1);
+        return System.currentTimeMillis() + Base64.encodeAsString(file.getName().getBytes()) + ext;
+    }
+
     private File convertToFile(MultipartFile multipartFile) {
-        File convertedFile = new File(multipartFile.getOriginalFilename());
+        File convertedFile = new File(Objects.requireNonNull(multipartFile.getOriginalFilename()));
         try (FileOutputStream fos = new FileOutputStream(convertedFile)) {
             fos.write(multipartFile.getBytes());
         } catch (IOException e) {
@@ -57,7 +72,7 @@ public class ImageService {
     }
 
     private boolean isDefault(String fileName) {
-        return defaultImageUrl.equals(fileName);
+        return defaultImage.equals(fileName);
     }
 }
